✅ VC2ZMZ tests: split pre-push full-fast guard
