+++ packages/agentplane/src/cli/run-cli.core.hooks.hook-run.test.ts
@@ -1,6 +1,5 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { execFile, execFileSync } from "node:child_process";
-import { readFileSync } from "node:fs";
 import {
   chmod,
   mkdir,
@@ -826,82 +825,6 @@ describe("runCli hooks run", { timeout: HOOKS_SUITE_TIMEOUT_MS }, () => {
     );
     expect(String(failure?.stderr ?? "")).not.toContain("pre-push blocked: ci:local:fast failed.");
     expect(await readFile(trackedFile, "utf8")).toBe("export const example = 2;\n");
-  });
-
-  it("pre-push hook fails fast when changed files require full-fast local CI", async () => {
-    const root = await mkGitRepoRoot();
-    await mkdir(path.join(root, ".agentplane", "tasks", "202601010101-ABCDEF"), {
-      recursive: true,
-    });
-    await mkdir(path.join(root, "scripts"), { recursive: true });
-    await writeFile(
-      path.join(root, ".agentplane", "tasks", "202601010101-ABCDEF", "README.md"),
-      "# Task\n",
-      "utf8",
-    );
-    await writeFile(
-      path.join(root, "package.json"),
-      JSON.stringify(
-        {
-          name: "hook-test",
-          private: true,
-          scripts: {
-            "format:check": "node scripts/format-check.mjs",
-            "ci:local:fast": "node scripts/ci-fast.mjs",
-          },
-        },
-        null,
-        2,
-      ),
-      "utf8",
-    );
-    await writeFile(
-      path.join(root, "scripts", "format-check.mjs"),
-      'import { writeFileSync } from "node:fs";\nwriteFileSync("format-ran.txt", "yes");\n',
-      "utf8",
-    );
-    await writeFile(
-      path.join(root, "scripts", "ci-fast.mjs"),
-      'import { writeFileSync } from "node:fs";\nwriteFileSync("ci-ran.txt", "yes");\n',
-      "utf8",
-    );
-    const execFileAsync = promisify(execFile);
-    await execFileAsync("git", ["add", ".agentplane", "package.json", "scripts"], {
-      cwd: root,
-    });
-    await execFileAsync("git", ["commit", "-m", "seed hook scripts"], { cwd: root });
-    const remoteSha = readFileSync(path.join(root, ".git", "refs", "heads", "main"), "utf8").trim();
-
-    const pkg = JSON.parse(await readFile(path.join(root, "package.json"), "utf8")) as {
-      description?: string;
-    };
-    pkg.description = "broad change";
-    await writeFile(path.join(root, "package.json"), `${JSON.stringify(pkg, null, 2)}\n`, "utf8");
-    await execFileAsync("git", ["add", "package.json"], { cwd: root });
-    await execFileAsync("git", ["commit", "-m", "✨ ABCDEF hooks: change package metadata"], {
-      cwd: root,
-    });
-    const localSha = readFileSync(path.join(root, ".git", "refs", "heads", "main"), "utf8").trim();
-
-    let failure: (Error & { stderr?: string | Buffer; stdout?: string | Buffer }) | null = null;
-    try {
-      execFileSync("node", [PRE_PUSH_HOOK_SCRIPT], {
-        cwd: root,
-        stdio: "pipe",
-        input: `refs/heads/main ${localSha} refs/heads/main ${remoteSha}\n`,
-      });
-    } catch (error) {
-      failure = error as Error & { stderr?: string | Buffer; stdout?: string | Buffer };
-    }
-
-    expect(failure).not.toBeNull();
-    expect(String(failure?.stderr ?? "")).toContain(
-      "pre-push blocked: changed files require the full-fast local CI lane",
-    );
-    expect(String(failure?.stderr ?? "")).toContain("selector=full-fast");
-    expect(String(failure?.stderr ?? "")).toContain("bun run ci:local:fast");
-    await expect(pathExists(path.join(root, "format-ran.txt"))).resolves.toBe(false);
-    await expect(pathExists(path.join(root, "ci-ran.txt"))).resolves.toBe(false);
   });
 
   it(
